+++ Fun.Blazor.Docs.Server/CustomElementsDemo.fs
@@ -1,16 +1,23 @@
 namespace Fun.Blazor.Docs.Server
 
+open Microsoft.AspNetCore.Components
 open Fun.Blazor
 open Fun.Htmx
 
 [<FunBlazorCustomElement>]
-type ServerDemoCounter() =
+type ServerDemoCounter() as this =
     inherit FunComponent()
 
     let mutable count = 0
 
+    [<Parameter>]
+    member val count_from_query = 0 with get, set
+
+    [<Parameter>]
+    member val count_from_form = 0 with get, set
+
     override _.Render() = div {
-        p { $"count = {count}" }
+        p { $"count = {count}; q {this.count_from_query}; f {this.count_from_form}" }
         button {
             onclick (fun _ -> count <- count + 1)
             "Click me"
@@ -58,9 +65,10 @@ type CustomElementsDemo =
                 p { "Integrate with htmx" }
                 button {
                     hxTrigger' hxEvt.mouse.click
-                    hxRequestCustomElement typeof<ServerDemoCounter> [
-                        "hi", "123"
-                    ]
+                    hxRequestCustomElement (QueryBuilder<ServerDemoCounter>()
+                        .Add((fun x -> x.count_from_query), 1)
+                        .Add((fun x -> x.count_from_form), 2)
+                    )
                     hxSwap_afterend
                     "Htmx click to add counter"
                 }
@@ -69,9 +77,11 @@ type CustomElementsDemo =
                 p { "Integrate SSR with htmx (the counter is not interactive)" }
                 button {
                     hxTrigger' hxEvt.mouse.click
-                    hxRequestBlazorSSR typeof<ServerDemoCounter> [
-                        "hi", "123"
-                    ]
+                    name (nameof Unchecked.defaultof<ServerDemoCounter>.count_from_form)
+                    value 3453453
+                    hxRequestBlazorSSR "post" (QueryBuilder<ServerDemoCounter>()
+                        .Add((fun x -> x.count_from_query), 3)
+                    )
                     hxSwap_afterend
                     "Htmx click to add counter"
                 }